Safari-609.1.11.1/LayoutTests/legacy-animation-engine/animations/computed-style-expected.txt

This tests computed style values from animation properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test1Style.webkitAnimationName is 'anim1'
PASS test2Style.webkitAnimationName is 'anim2, anim3'
PASS test3Style.webkitAnimationName is 'anim1, anim2, anim3'
PASS test1Style.webkitAnimationDuration is '10s'
PASS test2Style.webkitAnimationDuration is '5s, 2.5s'
PASS test1Style.webkitAnimationFillMode is 'backwards'
PASS test2Style.webkitAnimationFillMode is 'forwards, both'
PASS test1Style.webkitAnimationIterationCount is '10'
PASS test2Style.webkitAnimationIterationCount is '10, infinite'
PASS test1Style.webkitAnimationTimingFunction is 'linear'
PASS test2Style.webkitAnimationTimingFunction is 'linear, ease-in-out'
PASS test3Style.webkitAnimationTimingFunction is 'steps(1, start), steps(1, end), steps(5, end)'
PASS test1Style.webkitAnimationDirection is 'normal'
PASS test2Style.webkitAnimationDirection is 'normal, alternate'

PASS successfullyParsed is true

TEST COMPLETE
